docs: [WIP] Overhauled the basic docs for using Gladier
Note: These changes are dependent on the new directory structure above

This is a major pass over the existing docs in Gladier. It reorganizes
the docs in a more readable sequence, and splits up tutorial docs from
reference docs. Users first get familiar with Gladier through the
installation, setup, and getting started docs. The rest are reference
and should be fairly self-explanatory.

The examples have also been re-written to be more centric around using
the @generate_flow_definition style for composing flows.