docs: Added defaults to tar example, slight tweaks
Commented out 'funcx_endpoint_compute' so Gladier will catch the missing
value and complain. This will ensure the user sets that value. Otherwise,
the flow will start and result in a "FlowFailed". Useful, since people
tend to run things 'as is' and are confused by any errors which are not
extremely specific.