fix: Possible flows client 401 due to client caching
Glaider will cache the flows-client when it is created for performance
if many calls need to take place. However, this is a problem if a new
login added a new scope to the client. This change simply removes the
flows client cache on re-login, so it can be re-cached later.

This isn't a perfect fix, but it does fix a problem when registering
and running flows for the first time.