Prototype flow branching for tool chaining
Chaining works for complex flows that have arbitrary branches. This
doesn't yet support tool ailiases (duplication) due to how the old
system worked. This replaces the old system that relied on a linear
combination of flow states with a recursive implementation which can
handle branching.

The due to the linear nature of the old system, chaining flows together
was simplistic as there was only one option. The new system chains based
on a depth first search of the flow, whichever state comes first. This
still isn't the ideal option since flow authors may desire specific
termination points on any 'leaf' the flow ends. A simple solution for
automatically choosing these is to have tool authors list the transition
states on their tool if there is more than one state so the chaining
system can automatically link them.

More to come...