feat: Add "flow_transition_states" to BaseTools for determining Choice state
If a given flow uses Choice states and has more than one termination
state (End: True), flow_transition_states can be specified to direct
the built-in chaining class which states should continue to the next
tool in the `gladier_tools` list. States not in this list will result
in the flow ending.

This is only required for Tools with more than one End state. Flows
which do not use Choice (branching) may simply end the flow.