Pydantic based models for States for a flow
This provides a class-based method for defining flow states and
composing and chaining them together as desired. The pydantic based
classes retain compatibility with the Gladier Tool model, so a
`GladierBaseClient` can, for example, use these classes instead
of (and probably in addition to though that should probably be
discouraged) instances of `GladierBaseTool`.