Change flow generation
This puts the iteration through a next chain for flow generation in
the flow state rather than in the GladierFLow logic.

Note that this exposed a problem with flows that have loops as this
current method will recurse forever, but so does pydantic if you just
ask it to do something simple like printing a flow state where there
is a loop in the model.