Add fail/succeed builtins and update to compute step
The fail and succeed builtins are added as available step types.

Compute step is updated with some helpers for getting back the result
path for the output of the computation. Also, minor updates to allow
for proper over-riding of the `function_to_call` so that it can be
properly defined statically by subclasses of the compute step class.