Update to the GlobusComputeStep
This was not invoking the other managers, notably login and compute
properly. This seems to fix it, though one might find a more elegant
way of doing it.

Also, this includes an example that combines sub-classing
`GlobusComputeStep` to invoke a specific function and the use of a
`Choice` state. This example doesn't belong in this repo long term, so
should be removed prior to merging, but hopefully helps to evaluate
the PR to see how things work.