Provide a type variable for Globus Compute Functions
To set the function on a sub-class of the GlobusComputeStep, one must
actually provide the type definition in the sub-classes declaration of
the `function_to_call` property. As this is a rather complex/opaque
type, a type variable is provided in the module so that it can much
more easily be re-used/defined properly in new sub-classes of
`GlobusComputeStep` when binding the value for `function_to_call`.

Use of the variable is done in the ShellCmd tool.