Update behavior of the next method
This allows for more proper chaining of uses of the next() method such
as:

state = ActionState(...).next(ActionState(...)).next(ActionState(...))

where state will be the first state in the chain as a result.

Also, variations on the next() method are added for inserting a next
state after a referenced state and then retaining the rest of the flow
after or replacing the next part of the flow with the input parameter.