Update ParametrsInputPath to generate parameters by default
This introduces logic in the parameters generation for states which
derive from `ParaemtersOrInputPath` such that parameters from the
classes' model will be used to generate the parameters for the Action
statement. A sub-class can override `non_parameter_properties` to set
the class properties which should not be included in parameters, or
set `set_parameters_from_properties` to False to remove this behavior
entirely.