Change handling of precedence between input path and parameters
It turned out to be hard to specify input path on an action state
because paremters were also usually present thanks to default values
in the model of various action classes. And, the implemented policy
had been that if both input path and parameters were present, an error
would be thrown. This changes to that if input path is present, it
takes precedence and the parameters (including those from default
values) are ignored.