fix: Dependent scope change error after successful login
When the flows client raised a dependent scope change and successfully
logged in, there was still a problem of the cached scope change not being
cleared after login, which made the login manager think it still hadn't
been added. This is now properly cleared after login so new flows can
be started.