Change how next_state is initialized in StateWithNextOrEnd
Per discussion on #263, change how the `next_state` value is
initialized on `StateWithNextOrEnd`. There is a bit of verboseness to
the code to properly pass typing.

This probably conflicts with other outstanding PRs as I started this
branch from the current state of `main` which doesn't include #263 and
thus does not have the renaming of `GladierBaseState` in place.